requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
pydantic
python-dotenv
tqdm
yaml
pandas
scikit-learn
python-dotenv